+++ src/main/java/baseball/facade/GameFacade.java
@@ -1,4 +1,6 @@
-package baseball;
+package baseball.facade;
+
+import baseball.processor.*;
 
 import java.util.List;
