+++ src/main/java/baseball/processor/GameInitProcessorImpl.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.processor;
 
 import camp.nextstep.edu.missionutils.Console;
 import camp.nextstep.edu.missionutils.Randoms;
