+++ src/main/java/baseball/processor/NumberMatchProcessor.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.processor;
 
 import java.util.List;
