+++ src/test/java/baseball/NumberMatchProcessorTest.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.processor.NumberMatchProcessor;
+import baseball.processor.NumberMatchProcessorImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
