+++ src/test/java/baseball/NumberValidateTest.java
@@ -1,5 +1,6 @@
 package baseball;
 
+import baseball.processor.NumberValidate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
